chore: move fixtures to plugins
