+++ bentoml/testing/grpc/servicer.py
@@ -3,11 +3,15 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from bentoml.grpc.v1alpha1 import service_test_pb2 as pb
-from bentoml.grpc.v1alpha1 import service_test_pb2_grpc as services
-
 if TYPE_CHECKING:
     from grpc import aio
+
+    from bentoml.grpc.v1alpha1 import service_test_pb2 as pb
+    from bentoml.grpc.v1alpha1 import service_test_pb2_grpc as services
+else:
+    from bentoml.grpc.utils import import_generated_stubs
+
+    pb, services = import_generated_stubs(file="service_test.proto")
 
 
 class TestServiceServicer(services.TestServiceServicer):
