+++ tests/unit/_internal/utils/test_analytics.py
@@ -132,7 +132,7 @@ def test_track_serve_init(
     mock_usage_event_debugging: MagicMock,
     mock_do_not_track: MagicMock,
     mock_post: MagicMock,
-    noop_service: Service,
+    simple_service: Service,
     production: bool,
     caplog: LogCaptureFixture,
 ):
@@ -145,7 +145,7 @@ def test_track_serve_init(
     mock_response.text = "sent"
 
     analytics.usage_stats._track_serve_init(  # type: ignore (private warning)
-        noop_service,
+        simple_service,
         production=production,
         serve_info=analytics.usage_stats.get_serve_info(),
         serve_kind="http",
@@ -157,7 +157,7 @@ def test_track_serve_init(
     mock_usage_event_debugging.return_value = True
     with caplog.at_level(logging.INFO):
         analytics.usage_stats._track_serve_init(  # type: ignore (private warning)
-            noop_service,
+            simple_service,
             production=production,
             serve_info=analytics.usage_stats.get_serve_info(),
             serve_kind="http",
@@ -218,10 +218,12 @@ def test_filter_metrics_report(
 
 
 @patch("bentoml._internal.utils.analytics.usage_stats.do_not_track")
-def test_track_serve_do_not_track(mock_do_not_track: MagicMock, noop_service: Service):
+def test_track_serve_do_not_track(
+    mock_do_not_track: MagicMock, simple_service: Service
+):
     mock_do_not_track.return_value = True
     with analytics.track_serve(
-        noop_service,
+        simple_service,
         production=False,
         serve_info=analytics.usage_stats.get_serve_info(),
     ) as output:
@@ -236,7 +238,7 @@ def test_track_serve_do_not_track(mock_do_not_track: MagicMock, noop_service: Se
 def test_legacy_get_metrics_report(
     mock_prometheus_client: MagicMock,
     mock_do_not_track: MagicMock,
-    noop_service: Service,
+    simple_service: Service,
 ):
     mock_do_not_track.return_value = True
     mock_prometheus_client.multiproc.return_value = False
@@ -331,7 +333,7 @@ def test_track_serve(
     mock_track_serve_init: MagicMock,
     mock_post: MagicMock,
     mock_do_not_track: MagicMock,
-    noop_service: Service,
+    simple_service: Service,
     monkeypatch: MonkeyPatch,
     caplog: LogCaptureFixture,
 ):
@@ -344,7 +346,7 @@ def test_track_serve(
 
     with caplog.at_level(logging.INFO):
         with analytics.track_serve(
-            noop_service,
+            simple_service,
             production=False,
             metrics_client=mock_prometheus_client,
             serve_info=analytics.usage_stats.get_serve_info(),
